renamed concat to str-join to be more consistent with nushell terminology. Made work as an expression.
