+++ crates/nu_plugin_polars/src/dataframe/series/mod.rs
@@ -48,7 +48,6 @@ pub(crate) fn series_commands() -> Vec<Box<dyn PluginCommand<Plugin = PolarsPlug
         Box::new(ArgUnique),
         Box::new(AsDate),
         Box::new(AsDateTime),
-        Box::new(Concatenate),
         Box::new(Contains),
         Box::new(Cumulative),
         Box::new(GetDay),
@@ -74,6 +73,7 @@ pub(crate) fn series_commands() -> Vec<Box<dyn PluginCommand<Plugin = PolarsPlug
         Box::new(SetSeries),
         Box::new(SetWithIndex),
         Box::new(Shift),
+        Box::new(StrJoin),
         Box::new(StrLengths),
         Box::new(StrSlice),
         Box::new(StrFTime),
