+++ crates/nu_plugin_polars/src/dataframe/series/string/mod.rs
@@ -1,7 +1,7 @@
-mod concatenate;
 mod contains;
 mod replace;
 mod replace_all;
+mod str_join;
 mod str_lengths;
 mod str_slice;
 mod strftime;
@@ -10,10 +10,10 @@ mod to_integer;
 mod to_lowercase;
 mod to_uppercase;
 
-pub use concatenate::Concatenate;
 pub use contains::Contains;
 pub use replace::Replace;
 pub use replace_all::ReplaceAll;
+pub use str_join::StrJoin;
 pub use str_lengths::StrLengths;
 pub use str_slice::StrSlice;
 pub use strftime::StrFTime;
